% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data1.R
\docType{data}
\name{data1}
\alias{data1}
\title{A proteomics dataset with random protein expression values - no missing values.}
\format{
A matrix with 1045 rows and 30 variables.
}
\usage{
data1
}
\description{
A longitudinal proteomics dataset with five timepoints, two conditions and
three replicates for each sample at each timepoint in each condition. The expression
values for the random data have been generated using the \code{rnorm} function.
}
\keyword{datasets}
