% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{data2}
\alias{data2}
\title{A small dataset with 50 proteins and 30 samples for the example usage of RolDE}
\format{
A matrix with 50 rows and 30 variables.
}
\source{
\url{https://www.ebi.ac.uk/pride/archive/projects/PXD002099}
}
\usage{
data2
}
\description{
Based on the included semi-simulated UPS1 spike-in dataset (data3) with 50 randomly chosen proteins from the data.
}
\keyword{datasets}
