% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{data_descriptions}
\alias{data_descriptions}
\title{Short descriptions for the data sets included with the RolDE - package}
\format{
A vector with 4 elements.
}
\usage{
data_descriptions
}
\description{
Short descriptions for the datasets provided with the RolDE - package to test
the functionality of RolDE.
}
\keyword{datasets}
