% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des_matrix2.R
\docType{data}
\name{des_matrix2}
\alias{des_matrix2}
\title{A RolDE design matrix for data2}
\format{
A matrix with 30 rows and 4 variables.
}
\usage{
des_matrix2
}
\description{
A design matrix to be used with data2 for the RolDE-function.
}
\keyword{datasets}
