% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des_matrix4.R
\docType{data}
\name{des_matrix4}
\alias{des_matrix4}
\title{A RolDE design matrix for data4}
\format{
A matrix with 30 rows and 4 variables.
}
\usage{
des_matrix4
}
\description{
A design matrix to be used with data4 for the RolDE-function.
}
\keyword{datasets}
