% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des_matrix5.R
\docType{data}
\name{des_matrix5}
\alias{des_matrix5}
\title{A RolDE design matrix for data5}
\format{
A matrix with 189 rows and 4 variables.
}
\usage{
des_matrix5
}
\description{
A design matrix to be used with data5 for the RolDE-function.
}
\keyword{datasets}
