% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res1.R
\docType{data}
\name{res1}
\alias{res1}
\title{RolDE results for data1}
\format{
An object of class \code{list} of length 10.
}
\usage{
res1
}
\description{
RolDE results of data1 to be used for generating documentation only.
}
\keyword{datasets}
