% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res3.R
\docType{data}
\name{res3}
\alias{res3}
\title{RolDE results for data3}
\format{
An object of class \code{list} of length 10.
}
\usage{
res3
}
\description{
RolDE results of data3 to be used for generating documentation only.
}
\keyword{datasets}
