% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des_matrix1.R
\docType{data}
\name{des_matrix1}
\alias{des_matrix1}
\title{A RolDE design matrix for data1}
\format{
A matrix with 30 rows and 4 variables.
}
\usage{
des_matrix1
}
\description{
A design matrix to be used with data1 for the RolDE-function.
}
\keyword{datasets}
