% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des_matrix2.R
\docType{data}
\name{des_matrix2}
\alias{des_matrix2}
\title{A RolDE design matrix for data2}
\format{
A matrix with 30 rows and 4 variables.
}
\usage{
des_matrix2
}
\description{
A design matrix to be used with data2 for the RolDE-function.Column 1 contains sample names,
column 2 the condition information for each sample, column 3 indicates the timepoint for each sample,
and column 4 gives the Replicate or Individual each sample is coming from.
}
\keyword{datasets}
